+++ package.json
@@ -19,6 +19,8 @@
     "react-scripts": "5.0.0",
     "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
+    "redux-persist": "^6.0.0",
+    "redux-thunk": "^2.4.1",
     "sass": "^1.49.9",
     "web-vitals": "^2.1.4"
   },
