+++ src/store/index.js
@@ -1,16 +1,30 @@
-import { combineReducers, createStore } from "redux";
-import { conversationReducer} from "./conversation/reducer";
-import { messagesReducer } from "./mesages/reducer";
-
+import { conversationReducer } from "./conversation/reducer";
 import { profileReducer } from "./profile/reducer";
+import { combineReducers, createStore, applyMiddleware, compose } from "redux";
+import { persistReducer, persistStore } from "redux-persist";
+import storage from "redux-persist/lib/storage";
+import thunk from "redux-thunk";
+import { messagesReducer } from "./messages/reducer";
+
+const persistConfig = {
+  key: "gbMessenger",
+  storage,
+  whitelist: ["messages", "chats"],
+};
 
 const rootReducer = combineReducers({
   profile: profileReducer,
   chats: conversationReducer,
-  messages: messagesReducer
+  messages: messagesReducer,
 });
 
+const persistedReducer = persistReducer(persistConfig, rootReducer);
+
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+
 export const store = createStore(
-  rootReducer,
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  persistedReducer,
+  composeEnhancers(applyMiddleware(thunk))
 );
+
+export const persistor = persistStore(store);
