+++ src/store/messages/selectors.js
@@ -1,4 +1,3 @@
 export const selectMessages = (state) => state.messages;
-export const selectMessagesByChatId = (chatId) => (state) => {
+export const selectMessagesByChatId = (chatId) => (state) =>
   state.messages[chatId];
-};
